requirements.txt
streamlit==1.47.0
streamlit-markmap==1.0.1
google-generativeai==0.8.5
python-dotenv==1.1.1